+++ tests/strategies/test_reverse_fusion.py
@@ -70,6 +70,33 @@ def reverse_fusion_rules():
     yield FusionStrategy(col_idx=1, tracked=True)(t3).to_reverse_rule(0)
     yield FusionStrategy(row_idx=1, tracked=True)(t3.rotate270()).to_reverse_rule(0)
 
+    left_requirement = [GriddedPerm.point_perm((1, 0))]
+    right_requirement = [GriddedPerm.point_perm((2, 0))]
+    t4 = t.add_assumptions([left_overlap, left]).add_list_requirement(right_requirement)
+    yield FusionStrategy(col_idx=1, tracked=True)(t4).to_reverse_rule(0)
+    t5 = t.add_assumptions([left_overlap, left]).add_list_requirement(left_requirement)
+    yield FusionStrategy(col_idx=1, tracked=True)(t5).to_reverse_rule(0)
+
+    # # FOR A MORE EXHAUSTIVE SET OF TESTS UNCOMMENT THE FOLLOWING
+    # from itertools import chain, combinations
+    #
+    # for assumptions in chain.from_iterable(
+    #     combinations([left, right, left_overlap, right_overlap], i) for i in range(5)
+    # ):
+    #     for reqs in chain.from_iterable(
+    #         combinations([left_requirement, right_requirement], i) for i in range(2)
+    #     ):
+    #         tiling = t.add_assumptions(assumptions)
+    #         for req in reqs:
+    #             tiling = tiling.add_list_requirement(req)
+    #         rule = FusionStrategy(col_idx=1, tracked=True)(tiling)
+    #         rotate_tiling = tiling.rotate270()
+    #         rotate_rule = FusionStrategy(row_idx=1, tracked=True)(rotate_tiling)
+    #         yield rule
+    #         if left in assumptions or right in assumptions:
+    #             yield rule.to_reverse_rule(0)
+    #             yield rotate_rule.to_reverse_rule(0)
+
 
 @pytest.fixture
 def both_reverse_fusion_rule():
@@ -109,9 +136,6 @@ def test_positive_reverse_fusion(caplog):
     reverse_rule = rule.to_reverse_rule(0)
     logger.propagate = True
     with caplog.at_level(logging.WARNING):
-        reverse_rule.sanity_check(4)
-        assert len(caplog.records) == 2
-        assert (
-            "Skipping sanity checking counts" in caplog.text
-            and "Skipping sanity checking generation" in caplog.text
-        )
+        assert reverse_rule.sanity_check(4)
+        assert len(caplog.records) == 1
+        assert "Skipping sanity checking generation" in caplog.text